We remove all the static images in the vignette. Now every chunk is run for good in the Main Vignette. This is runing almost during 2 minutes. We use data subset and correct some mistakes that make the vignette runs for ages. (when symbol.col was set, part of code in dataprepare) We integrate SpatialExperiment. We can deal with that. Interoperability is better. We can also deal with SummarizedExperiment to extract counts. Counts pass to dataprepare can be from different classes. See dataprepare.
We have to do the differential-Vignette now.

Try to add signatures in the generic function.
Also symplify without testing for one example, it seems to work.

Reviews are here : https://github.com/Bioconductor/Contributions/issues/3554
What I have done is listed here :
https://docs.google.com/document/d/1HJiLlaTSmPpWt8oMdCc7ksM757VQ-fX8jIAm4jscMN4/edit?usp=sharing

Also we have done :

- Avoid overly verbose package startup messages in .onAttach.
- We add the infrasctucuture for Test.
- Avoid writing files for the user by default (filename argument in
learnParameters,BSRDataModel-method). This is set to NULL
- If finally removed the old way to print the figures in svg / pdf / png. Now you call the functions and user will handle the ouput.
That was a bad behavior we didn't correct.
- We remove png and dev.off calls .
- Rappdirs for package related directory was changed.
- Use of BioBase::testConnection was changed to curl::has_internet()
- One worker was set for the vignette.

Note : I d like to correct the way learnparameters print the pdf in order to include in the output in the vignette.

There is plenty of little things easier to do...

The big next step is to test the doc again (remove FALSE) without taking ages...