Dear reviewer,
We think we have followed most of your recommendations to fit best practices as possible. We hope we will satisfy bioconductor expectations now.
We try to resume what we have done in the next part of the text to help the reviewing process.
Thanks for your advice.
Kind regards.

BulkSignalR
We now consider interoperability with the Bioconductor matrix-type classes. Counts can be extracted from SpatialExperiment or SummarizedExperiment depending on the workflow in the input integration (see prepareDataset function in dataPrepare.R)
S4 generics have now explicit list of arguments. (Signature parameter added)
All examples are now runnable and NOT wrapped in if (FALSE). This has been a huge improvement in fact and a lot of work to make it work in a reasonable amount of time (R CMD check)..
Description
We changed rappdirs  to tools::R_user_dir("BulkSignalR", which="cache")
Namespace
We considered using more descriptive names (colA/B is now colClusterA/B …e.g., cmpName is comparisonName, etc). We removed unexplicit shortcut names and tried to give better definitions for function names as suggested.
We have considered the removal of Biobase::testBiocCConnection and we now use  curl::has_internet() function.
Vignettes
We Include package: BulkSignalR in the YAML front matter and remove the package version note in the Load packages section.
We removed any commented code from the vignette (BulkSignalR-Differential.Rmd)
We now only use one worker for the vignette.
We do not use static images anymore in the vignette. We use evaluated code to generate plots.
We rewrite the spatial workflow and now show a use case starting with a spatial experiment object from STexampleData and we mention VisiumIO.
We removed png and dev.off calls from the vignette and use only the functions in our package. We do not use static images anymore.
R
We do not use print in the example code for getComplexes
We avoid using get.
As mentioned before, we do not use if (FALSE) in examples anymore. BSRClusterComp-class.R).
We use more descriptive slot names for the BSRClusterComp class.
We now avoid conditionally defining generics and use setGeneric directly.
We use your own accessor and setter functions in learnParameters rather than obj@SLOT.
.getEmpiricalNull internal code is more user friendly and do not require extensive call to names that are present in obj@param extensive
For the allocate and fill method, we now create a list of finite length, e.g., in BSRDataModel-class.R: use vector("list", length(ligands(sig))) instead of vapply(x, length, integer(1))
We change the use of  as.BSRDataModelComp to the as method (SetAs) for S4 classes.

We did not avoid using a method to create a class and keep our approach in (defineClusterComp,BSRDataModelComp-method for creating a BSRClusterComp class; intialInference,BSRDataModelComp-method) but we are open on this point for discussion if it’s a real problem to do as we did.
We still want to keep the graphing functionality in the package.
We did not change the assign calls in the .onLoad function because it’s the only way we found to create env variables, in order to potentially modify them in another script outside BulkSignalR and pass them to another package we develop that will work with S4 classes from BulkSignalR .Very sincerely this is for me the most blocking point for everything you ask. These variables should be able to be called outside the library, be modified by the user, and these modifications should be taken into account in another dependent package we are developing.
We removed verbose package startup messages in .onAttach.
Tests
We set up the framework for unit testing with testthat for further development later. We have made a real effort to make examples runnable and also to execute the functions in the vignette in an acceptable time. We clearly do understand why but we also think that Unit testing should be incorporated in the flow of development and, in my opinion, it makes less sense to include them after.They will be added when bug correction is needed. That said, if it’s mandatory to have a minimum level of coverage, we will do it but we expect your leniency on this point : )

We tested the package under R Under development (unstable) (2024-11-03 r87286) with Bioc21. Rdevel CMD build run in 2:30 min and the Rdevel CMD check –no-build-vignettes in almost 10mins. We think this fits a normal behavior.